Backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.0
pydantic==2.9.2
sqlalchemy==2.0.29
psycopg[binary]==3.2.0
langchain==0.3.1
langchain-community==0.3.1
langchain-ollama==0.2.0
faiss-cpu==1.8.0
pdfplumber==0.11.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
firebase-admin==6.5.0
python-decouple==3.8